+++ apps/storybook/src/BottomSheetReact.tsx
@@ -1,5 +1,5 @@
 import {
-  BottomSheet,
+  BottomSheetCore,
   createBottomSheet,
   createPlaceholderBottomSheet,
   BOTTOM_SHEET_SELECTORS,
@@ -24,66 +24,67 @@ interface PlainBottomSheetProps {
 const placeholderBottomSheet = createPlaceholderBottomSheet();
 
 // TODO: Make it a React component library.
-export const BottomSheetReact = forwardRef<BottomSheet, PlainBottomSheetProps>(
-  function InnerBottomSheet(props, refToExpose) {
-    const [bottomSheet, setBottomSheet] = useState<BottomSheet>(
-      placeholderBottomSheet
-    );
-    const bottomSheetContentsWrapperRef = useRef<HTMLElement | null>(null);
+export const BottomSheetReact = forwardRef<
+  BottomSheetCore,
+  PlainBottomSheetProps
+>(function InnerBottomSheet(props, refToExpose) {
+  const [bottomSheet, setBottomSheet] = useState<BottomSheetCore>(
+    placeholderBottomSheet
+  );
+  const bottomSheetContentsWrapperRef = useRef<HTMLElement | null>(null);
 
-    useImperativeHandle(
-      refToExpose,
-      () => {
-        return bottomSheet;
-      },
-      [props, bottomSheet]
-    );
+  useImperativeHandle(
+    refToExpose,
+    () => {
+      return bottomSheet;
+    },
+    [props, bottomSheet]
+  );
 
-    useEffect(
-      function initiateBottomSheet() {
-        const mountingPoint =
-          props.mountingPoint === undefined || props.mountingPoint === null
-            ? window.document.body
-            : props.mountingPoint;
-        if (!mountingPoint) {
-          return;
-        }
-        if (bottomSheet.getIsMounted()) {
-          if (!bottomSheetContentsWrapperRef.current) {
-            return;
-          }
-
-          // TODO: Efficiently avoid rerendering/recreating of the bottom sheet
-          // while updating the user-provided content
-          // createPortal(props.children, bottomSheetContentsWrapperRef.current);
+  useEffect(
+    function initiateBottomSheet() {
+      const mountingPoint =
+        props.mountingPoint === undefined || props.mountingPoint === null
+          ? window.document.body
+          : props.mountingPoint;
+      if (!mountingPoint) {
+        return;
+      }
+      if (bottomSheet.getIsMounted()) {
+        if (!bottomSheetContentsWrapperRef.current) {
           return;
         }
 
-        // TODO: Initiate with user-provided props
-        const bottomSheetInstance = createBottomSheet({
-          content: props.options?.content ?? "",
-          ...props.options,
-        });
-        bottomSheetInstance.mount(mountingPoint);
+        // TODO: Efficiently avoid rerendering/recreating of the bottom sheet
+        // while updating the user-provided content
+        // createPortal(props.children, bottomSheetContentsWrapperRef.current);
+        return;
+      }
 
-        setBottomSheet(bottomSheetInstance);
+      // TODO: Initiate with user-provided props
+      const bottomSheetInstance = createBottomSheet({
+        content: props.options?.content ?? "",
+        ...props.options,
+      });
+      bottomSheetInstance.mount(mountingPoint);
 
-        bottomSheetContentsWrapperRef.current = document.querySelector(
-          BOTTOM_SHEET_SELECTORS.CONTENTS_WRAPPER
-        );
+      setBottomSheet(bottomSheetInstance);
 
-        return () => {
-          bottomSheet.unmount();
-        };
-      },
-      [props.mountingPoint, props.options, props.options?.content]
-    );
+      bottomSheetContentsWrapperRef.current = document.querySelector(
+        BOTTOM_SHEET_SELECTORS.CONTENTS_WRAPPER
+      );
 
-    return (
-      <>
-        {bottomSheetContentsWrapperRef.current &&
-          createPortal(props.children, bottomSheetContentsWrapperRef.current)}
-      </>
-    );
-  }
-);
+      return () => {
+        bottomSheet.unmount();
+      };
+    },
+    [props.mountingPoint, props.options, props.options?.content]
+  );
+
+  return (
+    <>
+      {bottomSheetContentsWrapperRef.current &&
+        createPortal(props.children, bottomSheetContentsWrapperRef.current)}
+    </>
+  );
+});