+++ apps/next-app/next.config.js
@@ -1,8 +1,0 @@
-/** @type {import('next').NextConfig} */
-const nextConfig = {
-  compiler: {
-    emotion: true,
-  },
-};
-
-module.exports = nextConfig;