+++ apps/next/app/data.ts
@@ -1,0 +1,153 @@
+import {
+  CurrencyDollarIcon,
+  GlobeAmericasIcon,
+} from "@heroicons/react/24/outline";
+import product0 from "../public/product-page-0.jpg";
+import product1 from "../public/product-page-1.jpg";
+import product2 from "../public/product-page-2.jpg";
+import relatedProduct0 from "../public/product-related-0.jpg";
+
+export const navigation = {
+  categories: [
+    {
+      id: "women",
+      name: "Women",
+    },
+    {
+      id: "men",
+      name: "Men",
+    },
+    {
+      id: "company",
+      name: "Company",
+    },
+    {
+      id: "stores",
+      name: "Stores",
+    },
+  ],
+};
+export const product = {
+  name: "Basic Tee",
+  price: "35 €",
+  href: "#",
+  breadcrumbs: [
+    { id: 1, name: "Women", href: "#" },
+    { id: 2, name: "Clothing", href: "#" },
+  ],
+  images: [
+    {
+      id: 1,
+      imageSrc: product0,
+      imageAlt: "Back of women's Basic Tee in black.",
+      primary: true,
+    },
+    {
+      id: 2,
+      imageSrc: product1,
+      imageAlt: "Side profile of women's Basic Tee in black.",
+      primary: false,
+    },
+    {
+      id: 3,
+      imageSrc: product2,
+      imageAlt: "Front of women's Basic Tee in black.",
+      primary: false,
+    },
+  ],
+  colors: [
+    { name: "Black", bgColor: "bg-gray-900", selectedColor: "ring-gray-900" },
+    {
+      name: "Heather Grey",
+      bgColor: "bg-gray-400",
+      selectedColor: "ring-gray-400",
+    },
+  ],
+  sizes: [
+    { name: "XXS", inStock: true },
+    { name: "XS", inStock: true },
+    { name: "S", inStock: true },
+    { name: "M", inStock: true },
+    { name: "L", inStock: true },
+    { name: "XL", inStock: false },
+  ],
+  description: `
+    <p>The Basic tee is an honest new take on a classic. The tee uses super soft, pre-shrunk cotton for true comfort and a dependable fit. They are hand cut and sewn locally, with a special dye technique that gives each tee it's own look.</p>
+    <p>Looking to stock your closet? The Basic tee also comes in a 3-pack or 5-pack at a bundle discount.</p>
+  `,
+  details: [
+    "Only the best materials",
+    "Ethically and locally made",
+    "Pre-washed and pre-shrunk",
+    "Machine wash cold with similar colors",
+  ],
+};
+export const policies = [
+  {
+    name: "International delivery",
+    icon: GlobeAmericasIcon,
+    description: "Get your order in 2 days",
+  },
+  {
+    name: "Loyalty rewards",
+    icon: CurrencyDollarIcon,
+    description: "Don't look at other tees",
+  },
+];
+export const reviews = {
+  average: 3.9,
+  totalCount: 512,
+  featured: [
+    {
+      id: 1,
+      title: "Can't say enough good things",
+      rating: 5,
+      content: `
+        <p>I was really pleased with the overall shopping experience. My order even included a little personal, handwritten note, which delighted me!</p>
+        <p>The product quality is amazing, it looks and feel even better than I had anticipated. Brilliant stuff! I would gladly recommend this store to my friends. And, now that I think of it... I actually have, many times!</p>
+      `,
+      author: "Risako M",
+      date: "May 16, 2021",
+      datetime: "2021-01-06",
+    },
+    // More reviews...
+  ],
+};
+export const relatedProducts = [
+  {
+    id: 1,
+    name: "Basic Tee",
+    href: "#",
+    imageSrc: relatedProduct0,
+    imageAlt: "Front of men's Basic Tee in white.",
+    price: "35 €",
+    color: "Aspen White",
+  },
+  // More products...
+];
+export const footerNavigation = {
+  products: [
+    { name: "Bags", href: "#" },
+    { name: "Tees", href: "#" },
+    { name: "Objects", href: "#" },
+    { name: "Home Goods", href: "#" },
+    { name: "Accessories", href: "#" },
+  ],
+  company: [
+    { name: "Who we are", href: "#" },
+    { name: "Sustainability", href: "#" },
+    { name: "Press", href: "#" },
+    { name: "Careers", href: "#" },
+    { name: "Terms & Conditions", href: "#" },
+    { name: "Privacy", href: "#" },
+  ],
+  customerService: [
+    { name: "Contact", href: "#" },
+    { name: "Shipping", href: "#" },
+    { name: "Returns", href: "#" },
+    { name: "Warranty", href: "#" },
+    { name: "Secure Payments", href: "#" },
+    { name: "FAQ", href: "#" },
+    { name: "Find a store", href: "#" },
+  ],
+};