+++ packages/core/__tests__/setup/prepare-bottom-sheet.ts
@@ -1,7 +1,9 @@
-import type { BottomSheet, BottomSheetProps } from "src";
+import type { BottomSheetCore, BottomSheetCoreProps } from "src";
 import { createBottomSheet } from "src";
 
-export function prepareOpenBottomSheet(props: BottomSheetProps): BottomSheet {
+export function prepareOpenBottomSheet(
+  props: BottomSheetCoreProps
+): BottomSheetCore {
   const bottomSheet = createBottomSheet(props);
 
   bottomSheet.mount();