+++ packages/react/CHANGELOG.md
@@ -45,5 +45,5 @@
 
 - improve reactivity of props
 - Updated dependencies
-  - pbs-utility@0.0.2
+  - @plainsheet/utility@0.0.2
   - @plainsheet/plain-bottom-sheet-core@0.4.2